split out mobile and desktop users page
