+++ web/src/components/DesktopUsers/DesktopUsers.css
@@ -1,15 +1,15 @@
-.circles-wrapper {
+.desktop-users-wrapper {
     display: flex;
     padding: 5em 2em 0 2em;
 }
 
 @media only screen and (max-width: 750px) {
-    .circles-wrapper {
+    .desktop-users-wrapper {
         padding: 0
     }
 }
 
-.circles-status {
+.desktop-users-status {
     padding: 13px;
     background-color: #ffffff;
     border-radius: 5px;
@@ -20,7 +20,7 @@
     max-width: 100%;
 }
 
-.circles-user-cards-container {
+.desktop-users-user-cards-container {
     padding: 0 50px;
     height: 550px;
     overflow-y: scroll;
@@ -36,19 +36,19 @@
 }
 
 /* Hide scrollbar for Chrome, Safari and Opera */
-.circles-user-cards-container::-webkit-scrollbar {
+.desktop-users-user-cards-container::-webkit-scrollbar {
     display: none;
 }
 
 /* Hide scrollbar for IE, Edge and Firefox */
-.circles-user-cards-container {
+.desktop-users-user-cards-container {
     -ms-overflow-style: none;  /* IE and Edge */
     scrollbar-width: none;  /* Firefox */
 }
 
 /* Mobile */
 @media only screen and (max-width: 750px) {
-    .circles-user-cards-container {
+    .desktop-users-user-cards-container {
         height: calc(100% - 220px);
         padding: 5px;
         top: 0;
@@ -58,8 +58,7 @@
     }
 }
 
-
-.circles-boards-wrapper {
+.desktop-users-boards-wrapper {
     overflow-x: scroll;
     overflow-y: hidden;
     direction:ltr;
@@ -69,8 +68,7 @@
     width: 100%;
 }
 
-.circles-boards {
+.desktop-users-boards {
     height: 100%;
     display: flex;
 }
-
