--- web/src/pages/Users/Users.css
@@ -1,113 +1,0 @@
-.users-wrapper {
-    display: flex;
-    padding: 5em 2em 0 2em;
-}
-
-@media only screen and (max-width: 750px) {
-    .users-wrapper {
-        padding: 0;
-    }
-}
-
-.users-status {
-    padding: 13px;
-    background-color: #ffffff;
-    border-radius: 5px;
-    box-shadow: 2px 2px 1px 1px #e0e0e0;
-    margin-bottom: 20px;
-    color: gray;
-    width: 100%;
-    max-width: 100%;
-}
-
-.users-grid-container {
-    display: grid;
-    width: 100%;
-    grid-template-columns: repeat(auto-fit, minmax(260px, 1fr));
-    grid-column-gap: 30px;
-    grid-row-gap: 30px;
-    padding: 5em 50px 0 50px;
-}
-
-.users-user-card-wrapper {
-    background-color: #ffffff;
-    width: 100%;
-    height: 100px;
-    border-radius: 5px;
-    box-shadow: 2px 2px 1px 1px #e0e0e0;
-    display: flex;
-    align-items: center;
-    cursor: pointer;
-}
-
-.users-user-card-right {
-    display: flex;
-    flex: 1;
-    align-items: center;
-    cursor: pointer;
-    justify-content: space-between;
-}
-
-.users-user-card-avatar {
-    height: 100px;
-    width: 100px;
-    flex-shrink: 0;
-}
-
-.users-user-card-avatar-img {
-    width: 100%;
-    height: 100%;
-}
-
-
-.users-user-card-name {
-    font-weight: bold;
-    font-size: large;
-    margin-left: 20px;
-    overflow: hidden;
-    white-space: nowrap;
-    text-overflow: ellipsis;
-}
-
-.users-user-card-join-time {
-    margin-left: 20px;
-    margin-top: 3px;
-    color: #767676;
-    font-size: small;
-}
-
-.users-user-card-follow-button {
-    width: 25px;
-    height: 25px;
-    margin-left: 2px;
-    margin-right: 5px;
-}
-
-.users-user-card-follow-button-disabled {
-    color: lightgray;
-    cursor: not-allowed;
-}
-
-/* Mobile */
-@media only screen and (max-width: 750px) {
-    .users-grid-container {
-        padding: 15px;
-        top: 0;
-        grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
-        grid-column-gap: 5px;
-        grid-row-gap: 5px;
-    }
-
-    .users-user-card-wrapper {
-        height: 50px;
-    }
-
-    .users-user-card-avatar {
-        height: 50px;
-        width: 50px;
-    }
-
-    .users-user-card-join-time {
-        display: none;
-    }
-}
